Finished MVP, used styled component to style the cards,Added an event handler to change the border color Onclick
